Refactor src structure
