+++ src/core/setupScene.js
